feat: add frontmatter support for pages
Add support for frontmatter parsing in pages. When a page is compiled,
it returns an object with any frontmatter properties, plus a
`contents` property containing the body of the file without the
frontmatter section.

Additionally, pages are stored under their filename minus the extension.

Also add esModuleInterop flag to this project.

BREAKING CHANGE: pages now compile to objects with a `contents` property
rather to strings